Migration to Firestore - Resources
Resources can now be added to topics

Kind of inefficient...but couldn't find a better way. Every resource is pulled manually upon opening a Topic. This could be avoided by keeping all of the information about that resource inside the Topic's subcollection, but that would require creative ways to keep data synced if changes would be made to the original resource.

Also cleaned up a bunch of code here as well.